scripts/requirements.txt
# Requirements for AI-Native Commerce Platform scripts
requests>=2.28.0
python-dotenv>=1.0.0
colorama>=0.4.6
pandas>=1.5.0
openpyxl>=3.1.0  # For Excel file support
langchain>=0.0.267
langchain_openai>=0.0.2
langchain_chroma>=0.0.1
